fix: Transform DocumentPublic Access from handler to Portlet - EXO-81854
Before this fix, documentPublicAccess use an handler.
As we change the behaviour with meeds-io/MIPs#203 and switch to classic REST portlet, we need to update this portlet.
