+++ documents-webapp/src/main/webapp/vue-app/download-document/main.js
@@ -47,14 +47,11 @@ const urls = [
   `/social/i18n/locale.portal.login?lang=${lang}`
 ];
 
-export function init(id, params) {
+export function init(id) {
   exoi18n.loadLanguageAsync(lang, urls).then(i18n => {
     // init Vue app when locale ressources are ready
     Vue.createApp({
-      data: {
-        params: params,
-      },
-      template: `<download-document id="${id}" :params="params" />`,
+      template: `<download-document id="${id}" />`,
       vuetify,
       i18n
     }, `#${id}`, 'Documents');
